+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Improved the "cannot craft! loop in processing..." error message (raoulvdberge)
 - Fixed error logs when toggling the Pattern Grid from and to processing mode (raoulvdberge)
 - Fixed pattern slots in Crafters not being accessible (raoulvdberge)
+- Fixed rare Grid crash (raoulvdberge)
 - Storage disk and block stored and capacity counts are formatted now in the tooltip (raoulvdberge)
 - Made the Disk Manipulator unsided (inserting goes to insert slots and extracting from output slots) (raoulvdberge)